refactor: Get rid of separate _meta module
