--- src/cleantest/_meta/__init__.py
@@ -1,10 +1,0 @@
-#!/usr/bin/env python3
-# Copyright 2023 Jason C. Nucciarone
-# See LICENSE file for licensing details.
-
-"""Private metaclasses used inside cleantest.
-
-Do not use these modules outside cleantest.
-"""
-
-from .base_configurer import BaseConfigurer
