--- src/cleantest/_meta/mixins/__init__.py
@@ -1,10 +1,0 @@
-#!/usr/bin/env python3
-# Copyright 2023 Jason C. Nucciarone
-# See LICENSE file for licensing details.
-
-"""Private mixins used inside cleantest.
-
-Do not use these modules outside cleantest.
-"""
-
-from .resettable import Resettable
